Implement Precision Iterator for real_rational class
The case of real_rational class is added to contructor of const_precision iterator.The iterator will constructed by making real_explicit or
real_operation instance of the real_rational and then making the iterator of that real_explicit or real_rational number.

The member "_real_ptr" of class const_precision_iterator is changed to shared pointer from raw pointer, and appropriate changes were made to contructor.

The contructor  of real_data is changed to initialize _precision_itr for real_rational also.