Bug Fix in Division operation
Division operation was showing "Division by zero" error when denomiator was a multiplication operation and initial iterations of that operation were giving zero as lower bound.
For example in this expression 2.4/(6.87*1.45), here (6.87*1.45) will give lower bound as zero for intial iteration because 1.45 will give [0, 2] in first iteration. Division should not throw any kind of error for such expressions.