Fix bug in trigonometric function
The system froze when calculating trigonometric functions of large expressions. Initially the lower and upper bounds of input values of trigonometric functions may be too large, the difference may be much greater than 2π. In such cases, there is no need to calculate sin and cos for those intervals and we simply need to iterate for more accurate input. So, calculating sin_cos only when difference is less than 8 fixed the issue.